Metal changes for internal review

+++ modules/javafx.graphics/src/main/java/com/sun/glass/ui/ios/IosView.java
@@ -60,7 +60,7 @@ final class IosView extends View {
     // See View
     @Override protected void _enableInputMethodEvents(long ptr, boolean enable) { }
 
-    @Override native protected int _getNativeFrameBuffer(long ptr);
+    @Override native protected long _getNativeFrameBuffer(long ptr);
     @Override native protected long _create(Map caps);
 
     @Override native protected long _getNativeView(long ptr);
