+++ modules/javafx.graphics/src/main/java/com/sun/prism/es2/ES2SwapChain.java
@@ -50,7 +50,7 @@ class ES2SwapChain implements ES2RenderTarget, Presentable, GraphicsResource {
     private float pixelScaleFactorY;
     // a value of zero corresponds to the windowing system-provided
     // framebuffer object
-    int nativeDestHandle = 0;
+    long nativeDestHandle = 0;
     private final boolean msaa;
     /**
      * An offscreen surface that acts as a persistent backbuffer, currently
@@ -237,7 +237,7 @@ class ES2SwapChain implements ES2RenderTarget, Presentable, GraphicsResource {
 
     @Override
     public int getFboID() {
-        return nativeDestHandle;
+        return (int)nativeDestHandle;
     }
 
     @Override
